+++ src/lib/Components/Inbox/Conversations/index.tsx
@@ -25,6 +25,7 @@ interface State {
   dataSource: ListViewDataSource | null
   fetchingNextPage: boolean
   completed: boolean
+  initialLoadDone: boolean
 }
 
 export class Conversations extends React.Component<Props, State> {
@@ -39,6 +40,7 @@ export class Conversations extends React.Component<Props, State> {
       dataSource,
       completed: false,
       fetchingNextPage: false,
+      initialLoadDone: false,
     }
   }
 
@@ -85,7 +87,7 @@ export class Conversations extends React.Component<Props, State> {
   }
 
   fetchData(nextPage: boolean = true) {
-    if (this.state.fetchingNextPage || this.state.completed) {
+    if (this.state.fetchingNextPage) {
       return
     }
     const totalSize = this.props.relay.variables.totalSize + (nextPage ? PageSize : 0)
@@ -95,6 +97,7 @@ export class Conversations extends React.Component<Props, State> {
       if (readyState.done) {
         this.setState({
           fetchingNextPage: false,
+          initialLoadDone: true,
           dataSource: this.state.dataSource.cloneWithRows(this.conversations),
         })
 
@@ -112,8 +115,12 @@ export class Conversations extends React.Component<Props, State> {
         initialListSize={10}
         scrollEventThrottle={10}
         onEndReachedThreshold={10}
+        enableEmptySections={true}
         refreshControl={
-          <RefreshControl refreshing={this.state.fetchingNextPage} onRefresh={() => this.fetchData(false)} />
+          <RefreshControl
+            refreshing={this.state.fetchingNextPage && this.state.initialLoadDone}
+            onRefresh={() => this.fetchData(false)}
+          />
         }
         renderHeader={() => {
           return (